Conflicts:
	libopflex/cwrapper/offramework.cpp
	libopflex/cwrapper/ofloghandler.cpp
	libopflex/cwrapper/ofobjectlistener.cpp
	libopflex/cwrapper/test/cwrapper_test.cpp
	libopflex/engine/OpflexClientConnection.cpp
	libopflex/include/opflex/c/offramework_c.h
	libopflex/include/opflex/c/ofloghandler_c.h

+++ libopflex/comms/active_connection.cpp
@@ -6,6 +6,12 @@
  * and is available at http://www.eclipse.org/legal/epl-v10.html
  */
 
+/* This must be included before anything else */
+#if HAVE_CONFIG_H
+#  include <config.h>
+#endif
+
+
 #include <opflex/logging/internal/logging.hpp>
 #include <yajr/internal/comms.hpp>
 
@@ -45,7 +51,11 @@
         UvLoopSelector uvLoopSelector
     ) {
 
-    LOG(INFO) << host << ":" << service;
+    LOG(INFO)
+        << host
+        << ":"
+        << service
+    ;
 
     ::yajr::comms::internal::ActivePeer * peer;
     if (!(peer = new (std::nothrow) ::yajr::comms::internal::ActivePeer(
@@ -54,17 +64,23 @@
                     connectionHandler,
                     data,
                     uvLoopSelector))) {
-        LOG(WARNING) << ": out of memory, dropping new peer on the floor";
+        LOG(WARNING)
+            << ": out of memory, dropping new peer on the floor"
+        ;
         return NULL;
     }
 
-    LOG(DEBUG) << peer << " queued up for resolution";
+    VLOG(1)
+        << peer
+        << " queued up for resolution"
+    ;
     peer->insert(::yajr::comms::internal::Peer::LoopData::TO_RESOLVE);
 
     return peer;
 }
 
-namespace yajr { namespace comms {
+namespace yajr {
+    namespace comms {
 
 using namespace yajr::comms::internal;
 
@@ -72,11 +88,26 @@ void ::yajr::comms::internal::ActivePeer::retry() {
 
     if (destroying_) {
 
-        LOG(INFO) << this << "Not retrying because of pending destroy";
+        LOG(INFO)
+            << this
+            << "Not retrying because of pending destroy"
+        ;
 
         return;
 
     }
+    if (uvRefCnt_ != 1) {
+        LOG(INFO)
+            << this
+            << " has to wait for reference count to fall to 1 before retrying"
+        ;
+
+        insert(internal::Peer::LoopData::RETRY_TO_CONNECT);
+
+        return;
+
+    }
+
 
     struct addrinfo const hints = (struct addrinfo){
         /* .ai_flags    = */ 0,
@@ -102,12 +133,18 @@ void ::yajr::comms::internal::ActivePeer::retry() {
                     getHostname(),
                     getService(),
                     &hints))) {
-        LOG(WARNING) << "uv_getaddrinfo: [" << uv_err_name(rc) << "] " <<
-            uv_strerror(rc);
+        LOG(WARNING)
+            << "uv_getaddrinfo: ["
+            << uv_err_name(rc)
+            << "] "
+            << uv_strerror(rc)
+        ;
         onError(rc);
         insert(internal::Peer::LoopData::RETRY_TO_CONNECT);
     } else {
-        LOG(DEBUG) << this << " up() for a pending getaddrinfo()";
+        VLOG(1)
+            << this
+            << " up() for a pending getaddrinfo()";
         up();
         insert(internal::Peer::LoopData::ATTEMPTING_TO_CONNECT);
     }
@@ -131,10 +168,29 @@ void on_active_connection(uv_connect_t *req, int status) {
 
     ActivePeer * peer = Peer::get(req);  // can't possibly crash yet
 
-    LOG(DEBUG) << peer;
+    if (status == UV_ECANCELED) {
+
+        /* the peer might have been deleted, so we have to avoid accessing any
+         * of its members */
+        VLOG(1)
+            << peer
+            << " has had a connection attempt canceled"
+        ;
+
+        peer->onError(status);
+
+        return;
+    }
+
+    VLOG(1)
+        << peer;
 
     if (peer->destroying_) {
-        LOG(INFO) << peer << " peer is being destroyed. down() it";
+        LOG(INFO)
+            << peer
+            << " peer is being destroyed. down() it"
+            << " NOT!"
+        ;
         peer->down();
         return;
     }
@@ -143,20 +199,28 @@ void on_active_connection(uv_connect_t *req, int status) {
 
     int rc;
     if (status < 0) {
-        LOG(WARNING) << "connect: [" << uv_err_name(status) <<
-            "] " << uv_strerror(status);
+        LOG(WARNING)
+            << "connect: ["
+            << uv_err_name(status)
+            << "] "
+            << uv_strerror(status)
+        ;
         if ((rc = connect_to_next_address(peer))) {
-            LOG(WARNING) << "connect: no more resolved addresses";
+            LOG(WARNING)
+                << peer
+                << "connect: no more resolved addresses"
+            ;
 
             /* since all attempts for all possible IP addresses have failed,
              * issue an error callback
              */
             peer->onError(rc);
 
-            /* retry later */
             if (!uv_is_closing((uv_handle_t*)&peer->handle_)) {
-                uv_close((uv_handle_t*)&peer->handle_, NULL);
+                uv_close((uv_handle_t*)&peer->handle_, on_close);
             }
+
+            /* retry later */
             retry_later(peer);
         }
         return;
@@ -186,7 +250,7 @@ void debug_resolution_entries(struct addrinfo const * ai);
 
 void on_resolved(uv_getaddrinfo_t * req, int status, struct addrinfo *resp) {
 
-    LOG(DEBUG);
+    VLOG(2);
 
     ActivePeer * peer = Peer::get(req);
     assert(!peer->passive_);
@@ -195,17 +259,25 @@ void on_resolved(uv_getaddrinfo_t * req, int status, struct addrinfo *resp) {
     void retry_later(ActivePeer * peer);
 
     if (peer->destroying_) {
-        LOG(INFO) << peer << " peer is being destroyed. down() it";
+        LOG(INFO)
+            << peer
+            << " peer is being destroyed. down() it"
+        ;
         peer->down();
         return;
     }
 
     if (status < 0) {
-        LOG(WARNING) << "getaddrinfo callback error: [" << uv_err_name(status) <<
-            "] " << uv_strerror(status);
+        LOG(WARNING)
+            << "getaddrinfo callback error: ["
+            << uv_err_name(status)
+            << "] "
+            << uv_strerror(status)
+        ;
         peer->status_ = Peer::kPS_FAILED_TO_RESOLVE;
         uv_freeaddrinfo(resp);
 
+        peer->down();
         return retry_later(peer);
     }
 
@@ -215,24 +287,33 @@ void on_resolved(uv_getaddrinfo_t * req, int status, struct addrinfo *resp) {
 
     int rc;
 
-    /* FIXME: pass the loop along */
     if ((rc = peer->tcpInit())) {
+
+        peer->down();
         return retry_later(peer);
+
     }
 
     peer->_.ai_next = peer->_.ai = resp;
     if ((rc = connect_to_next_address(peer))) {
-        LOG(WARNING) << "connect_to_next_address: [" << uv_err_name(rc) << "] " <<
-            uv_strerror(rc);
+        LOG(WARNING)
+            << "connect_to_next_address: ["
+            << uv_err_name(rc)
+            << "] "
+            << uv_strerror(rc)
+        ;
         if (!uv_is_closing((uv_handle_t*)&peer->handle_)) {
-            uv_close((uv_handle_t*)&peer->handle_, NULL);
+            uv_close((uv_handle_t*)&peer->handle_, on_close);
         }
         return retry_later(peer);
     }
 
     peer->status_ = Peer::kPS_CONNECTING;
 
-    LOG(DEBUG) << peer << " waiting for connection completion";
+    VLOG(1)
+        << peer
+        << " waiting for connection completion"
+    ;
 
 }
 
@@ -250,7 +331,7 @@ void on_resolved(uv_getaddrinfo_t * req, int status, struct addrinfo *resp) {
 
 void debug_address(struct addrinfo const * ai, size_t m = 0) {
 
-    if (!LOG_SHOULD_EMIT(DEBUG)) {
+    if (!VLOG_IS_ON(3)) {
         return;
     }
 
@@ -283,34 +364,30 @@ void debug_address(struct addrinfo const * ai, size_t m = 0) {
                        NI_NUMERICHOST | NI_NUMERICSERV
                       );
 
-    LOG(DEBUG)
-        << msg[m][0] << host
-        << msg[m][1] << service
-    ;
-    LOG(DEBUG)
-        << " ai_flags=" << ai->ai_flags
-    ;
-    LOG(DEBUG)
-        << " ai_family=" << ai->ai_family
-    ;
-    LOG(DEBUG)
-        << " ai_socktype=" << ai->ai_socktype
-    ;
-    LOG(DEBUG)
-        << " ai_protocol=" << ai->ai_protocol
-    ;
-    LOG(DEBUG)
-        << " ai_addrlen=" << ai->ai_addrlen
-    ;
-    LOG(DEBUG)
-        << " ai_canonname=" << (ai->ai_canonname ?: "")
+    VLOG(4)
+        << msg[m][0]
+        <<           host
+        << msg[m][1]
+        <<           service
+        <<   " ai_flags="
+        << ai->ai_flags
+        <<   " ai_family="
+        << ai->ai_family
+        <<   " ai_socktype="
+        << ai->ai_socktype
+        <<   " ai_protocol="
+        << ai->ai_protocol
+        <<   " ai_addrlen="
+        << ai->ai_addrlen
+        <<    " ai_canonname="
+        << (ai->ai_canonname ?: "")
     ;
 
 }
 
 void debug_resolution_entries(struct addrinfo const * ai) {
 
-    if (!LOG_SHOULD_EMIT(DEBUG)) {
+    if (!VLOG_IS_ON(3)) {
         return;
     }
 
@@ -322,13 +399,30 @@ void debug_resolution_entries(struct addrinfo const * ai) {
 
 void retry_later(ActivePeer * peer) {
 
-    LOG(DEBUG) << peer;
+    if (peer->destroying_) {
+        LOG(INFO)
+            << peer
+            << " peer is being destroyed. not inserting in RETRY_TO_CONNECT"
+        ;
+
+        return;
+
+    }
+
+    VLOG(1)
+        << peer
+    ;
 
     peer->unlink();
     peer->insert(internal::Peer::LoopData::RETRY_TO_CONNECT);
 
-    LOG(DEBUG) << peer << " down() for a retry_later()";
-    peer->down();
+ // VLOG(1)
+ //     << peer
+ //     << " down() for a retry_later()"
+ // ;
+
+    assert(peer->uvRefCnt_ > 0);
+ // peer->down();
 
     return;
 }
@@ -337,7 +431,9 @@ void swap_stack_on_close(uv_handle_t * h) {
 
     ActivePeer * peer = Peer::get<ActivePeer>(h);  // can't possibly crash yet
 
-    LOG(DEBUG) << peer;
+    VLOG(1)
+        << peer
+    ;
 
  // LOG(DEBUG) << "{" << peer << "}A down() for a swap_stack_on_close()";
  // peer->down();
@@ -346,16 +442,21 @@ void swap_stack_on_close(uv_handle_t * h) {
     /* FIXME: pass the loop along */
     if ((rc = peer->tcpInit())) {
 
+        peer->down();
         retry_later(peer);
 
         return;
     }
 
     if ((rc = connect_to_next_address(peer, false))) {
-        LOG(WARNING) << "connect_to_next_address: [" << uv_err_name(rc) << "] " <<
-            uv_strerror(rc);
+        LOG(WARNING)
+            << "connect_to_next_address: ["
+            << uv_err_name(rc)
+            << "] "
+            << uv_strerror(rc)
+        ;
         if (!uv_is_closing((uv_handle_t*)&peer->handle_)) {
-            uv_close((uv_handle_t*)&peer->handle_, NULL);
+            uv_close((uv_handle_t*)&peer->handle_, on_close);
         }
         return retry_later(peer);
     }
@@ -364,7 +465,9 @@ void swap_stack_on_close(uv_handle_t * h) {
 
 int connect_to_next_address(ActivePeer * peer, bool swap_stack) {
 
-    LOG(DEBUG) << peer;
+    VLOG(1)
+        << peer
+    ;
 
     struct addrinfo const * ai = peer->_.ai_next;
 
@@ -374,8 +477,12 @@ int connect_to_next_address(ActivePeer * peer, bool swap_stack) {
 
     while (ai && (rc = uv_tcp_connect(&peer->connect_req_, &peer->handle_,
                 ai->ai_addr, on_active_connection))) {
-        LOG(ai ? INFO : WARNING) << "uv_tcp_connect: [" << uv_err_name(rc) <<
-            "] " << uv_strerror(rc);
+        LOG(ai ? INFO : WARNING)
+            << "uv_tcp_connect: ["
+            << uv_err_name(rc)
+            << "] "
+            << uv_strerror(rc)
+        ;
 
         if (swap_stack) {
 
@@ -387,7 +494,9 @@ int connect_to_next_address(ActivePeer * peer, bool swap_stack) {
                 case -EAFNOSUPPORT:
                 case -EPROTOTYPE:
                 case -EINVAL:
-                    LOG(INFO) << "destroying socket and retrying";
+                    LOG(INFO)
+                        << "destroying socket and retrying"
+                    ;
                     if (!uv_is_closing((uv_handle_t*)&peer->handle_)) {
                         uv_close((uv_handle_t*)&peer->handle_, swap_stack_on_close);
                     }
@@ -410,14 +519,21 @@ int connect_to_next_address(ActivePeer * peer, bool swap_stack) {
     }
 
     if (rc) {
-        LOG(DEBUG) << "unable to issue a(nother) connect request";
+        VLOG(1)
+            << peer
+            << " unable to issue a(nother) connect request"
+        ;
     } else {
-        LOG(DEBUG) << "issued a connect request";
+        VLOG(1)
+            << peer
+            << " issued a connect request"
+        ;
     }
 
     return rc;
 }
 
 } /* yajr::comms::internal namespace */
+} /* yajr::comms namespace */
+} /* yajr namespace */
 
-}} /* yajr::comms and yajr namespaces */
